requirements.txt
discord.py
Flask
aiohttp
python-dotenv
dropbox
tenacity
